Initial changes for solaris support

+++ gcl/h/object.h
@@ -864,9 +864,9 @@ int FIXtemp;
 #endif
 
 
-#define	isUpper(xxx)	(((xxx)&0200) == 0 && isupper(xxx))
-#define	isLower(xxx)	(((xxx)&0200) == 0 && islower(xxx))
-#define	isDigit(xxx)	(((xxx)&0200) == 0 && isdigit(xxx))
+#define	isUpper(xxx)	(((xxx)&0200) == 0 && isupper((int)xxx))
+#define	isLower(xxx)	(((xxx)&0200) == 0 && islower((int)xxx))
+#define	isDigit(xxx)	(((xxx)&0200) == 0 && isdigit((int)xxx))
 enum ftype {f_object,f_fixnum};
 EXTER 
 char *alloca_val;
